+++ plugin/src/main/java/com/stratio/cassandra/lucene/search/Search.java
@@ -52,13 +52,13 @@ public class Search {
     private final Boolean refresh;
 
     /**
-     * Returns a new {@link Search} composed by the specified querying and filtering conditions.
+     * Constructor using the specified querying, filtering, sorting and refresh options.
      *
-     * @param query The {@link Condition} for querying, maybe {@code null} meaning no querying.
-     * @param filter The {@link Condition} for filtering, maybe {@code null} meaning no filtering.
-     * @param sort The {@link Sort} for the query. Note that is the order in which the data will be read before
-     * querying, not the order of the results after querying.
-     * @param refresh If this search must refresh the index before reading it.
+     * @param query the condition for querying, maybe {@code null} meaning no querying
+     * @param filter the condition for filtering, maybe {@code null} meaning no filtering
+     * @param sort the sort for the query. Note that is the order in which the data will be read before
+     * querying, not the order of the results after querying
+     * @param refresh if this search must refresh the index before reading it
      */
     public Search(Condition query, Condition filter, Sort sort, Boolean refresh) {
         this.query = query;
@@ -72,16 +72,16 @@ public class Search {
     }
 
     /**
-     * Returns {@code true} if this search requires full ranges scan, {code null} otherwise.
+     * Returns if this search requires full ranges scan.
      *
-     * @return {@code true} if this search requires full ranges scan, {code null} otherwise.
+     * @return {@code true} if this search requires full ranges scan, {code null} otherwise
      */
     public boolean requiresFullScan() {
         return usesRelevance() || usesSorting() || refresh && isEmpty();
     }
 
     /**
-     * Returns {@code true} if this search uses Lucene relevance formula, {@code false} otherwise.
+     * Returns if this search uses Lucene relevance formula.
      *
      * @return {@code true} if this search uses Lucene relevance formula, {@code false} otherwise.
      */
@@ -90,18 +90,18 @@ public class Search {
     }
 
     /**
-     * Returns {@code true} if this search uses field sorting, {@code false} otherwise.
+     * Returns if this search uses field sorting.
      *
-     * @return {@code true} if this search uses field sorting, {@code false} otherwise.
+     * @return {@code true} if this search uses field sorting, {@code false} otherwise
      */
     public boolean usesSorting() {
         return sort != null;
     }
 
     /**
-     * Returns {@code true} if this search doesn't specify any filter, query or sort, {@code false} otherwise.
+     * Returns if this search doesn't specify any filter, query or sort.
      *
-     * @return {@code true} if this search doesn't specify any filter, query or sort, {@code false} otherwise.
+     * @return {@code true} if this search doesn't specify any filter, query or sort, {@code false} otherwise
      */
     public boolean isEmpty() {
         return query == null && filter == null && sort == null;
@@ -110,7 +110,7 @@ public class Search {
     /**
      * Returns the field sorting to be used, maybe {@code null} meaning no field sorting.
      *
-     * @return The field sorting to be used, maybe {@code null} meaning no field sorting.
+     * @return the Lucene's sort
      */
     public Sort getSort() {
         return this.sort;
@@ -130,7 +130,7 @@ public class Search {
      * meaning no sorting.
      *
      * @param schema the {@link Schema}
-     * @return The Lucene {@link SortField}s represented by this using {@code schema}.
+     * @return the Lucene {@link SortField}s represented by this using {@code schema}
      */
     public List<SortField> sortFields(Schema schema) {
         return sort == null ? null : sort.sortFields(schema);
@@ -159,6 +159,12 @@ public class Search {
         return booleanQuery.clauses().isEmpty() ? new MatchAllDocsQuery() : booleanQuery;
     }
 
+    /**
+     * Returns the Lucene {@link Query} represented by this using the specified {@link Schema}.
+     *
+     * @param schema the {@link Schema}
+     * @return a Lucene {@link Query}
+     */
     public Query query(Schema schema) {
         return query(schema, null);
     }
@@ -166,7 +172,7 @@ public class Search {
     /**
      * Validates this {@link Search} against the specified {@link Schema}.
      *
-     * @param schema A {@link Schema}.
+     * @param schema a {@link Schema}
      */
     public void validate(Schema schema) {
         if (query != null) {
@@ -187,6 +193,7 @@ public class Search {
                           .add("query", query)
                           .add("filter", filter)
                           .add("sort", sort)
+                          .add("refresh", refresh)
                           .toString();
     }
 }