+++ plugin/src/test/java/com/stratio/cassandra/lucene/search/SearchTest.java
@@ -115,11 +115,13 @@ public class SearchTest {
     public void testToString() {
         Search search = search().query(match("field", "value"))
                                 .filter(match("field", "value"))
-                                .sort(field("field")).build();
+                                .sort(field("field"))
+                                .refresh(true).build();
         assertEquals("Method #toString is wrong",
                      "Search{query=MatchCondition{boost=1.0, field=field, value=value}, " +
                      "filter=MatchCondition{boost=1.0, field=field, value=value}, " +
-                     "sort=Sort{sortFields=[SimpleSortField{field=field, reverse=false}]}}",
+                     "sort=Sort{sortFields=[SimpleSortField{field=field, reverse=false}]}, " +
+                     "refresh=true}",
                      search.toString());
     }